Adds User Authentication
For the Dashboard to work, users need to be able to do things without
going through the bot. To that end, I've updated the authentication
methods to allow for server admins to use the API.

We check this by making a call out to Discord to figure out what servers
the user is on, and checking if the server in question is one they're an
admin on.