Dockerizing the API (#23)
* Merges Startup scripts

Not sure why I did things like this to begin with. Now instead of having
to select whether we're in prod or debug everytime we launch, we have a
single startup script which selects a mode based off an environment
variable.

* Changes from config file to env vars

Prereq for dockerizing, and just generally handier. Now all config
settings should be handled by environment variables on the machine
instead of as part of a config file.

* Dockerizes the server

Ayo we've got a way to build an image now.

* Sets up test envrionments

Important to actually be able to test things now that we're
containerized. A test context can be started using `make tester`, with
tests run by `make test` afterwards. The connection is hot, so no need
to reload

Co-authored-by: matthew-robertson <m33rober@uwaterloo.ca>