Fixes prod mode
The check to see if we were in dev mode was incorrect. It is now
considered to be in development mode if the `DEVELOPMENT_MODE`
environment variable is present.

Also "fixes" an issue where tasks would get queued up and cause the
bot to die, by making the thread_count an environment variable.